requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.0.5
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
